Refactored folder structure for course placement source code files
